pyproject.toml
[project]
name = "ml-exam"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen-agentchat>=0.7.5",
    "autogen-ext[chromadb,ollama,openai]>=0.7.5",
    "chonkie[chroma,st,gemini,semantic]>=1.4.2",
    "dotenv>=0.9.9",
    "pymarkdownlnt>=0.9.33",
]
